requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Azure OpenAI
openai>=1.30.0

# Azure Speech (optional — voice features)
azure-cognitiveservices-speech>=1.38.0

# RAG / Vector Search
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
tiktoken>=0.7.0

# Audio processing (voice features)
pydub>=0.25.1

# Utilities
aiofiles>=23.0.0
httpx>=0.27.0
langchain-text-splitters